refactor #9: MemberRepository 패키지 이동
