+++ src/test/java/server/preonboarding/budgetmanager/service/AuthServiceMockTest.java
@@ -5,18 +5,16 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.BDDMockito;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.security.crypto.password.PasswordEncoder;
-import server.preonboarding.budgetmanager.MemberRepository;
 import server.preonboarding.budgetmanager.domain.Member;
 import server.preonboarding.budgetmanager.dto.AuthLoginRequest;
 import server.preonboarding.budgetmanager.dto.AuthLoginResponse;
 import server.preonboarding.budgetmanager.exception.CustomException;
 import server.preonboarding.budgetmanager.exception.ErrorCode;
+import server.preonboarding.budgetmanager.repository.MemberRepository;
 import server.preonboarding.budgetmanager.security.JwtProvider;
 
 import java.util.Optional;
